Added LWJGL
Added LWJGL and merging.

+++ src/main/java/com/patetlex/displayphoenix/file/Data.java
@@ -11,7 +11,9 @@ import com.patetlex.displayphoenix.util.StringHelper;
 
 import java.io.*;
 import java.nio.file.Files;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -24,6 +26,8 @@ public class Data {
     private static boolean created;
     private static Map<String, Object> data;
 
+    private static List<String> markedStorage = new ArrayList<>();
+
     /**
      * Store a JSON parsable object
      *
@@ -31,7 +35,8 @@ public class Data {
      * @param value  Object value
      */
     public static void store(String key, Object value) {
-        data.put(key, value);
+        if (value != null)
+            data.put(key, value);
     }
 
     /**
@@ -67,15 +72,19 @@ public class Data {
      */
     public static void clearCache() {
         File directory = new File(System.getProperty("user.home") + "/." + StringHelper.id(Application.getTitle()));
-        for (File subFile : directory.listFiles()) {
-            if (!subFile.isDirectory()) {
-                subFile.delete();
-            }
-        }
+        File data = new File(directory.getPath() + "/data.json");
+        data.delete();
+
         FileHelper.forEachSubFile(new File(directory.getPath() + "\\cache\\"), new FileIteration() {
             @Override
             public void iterate(File file) {
-                file.delete();
+                boolean flag = false;
+                for (String filePath : markedStorage) {
+                    if (file.getPath().contains(filePath))
+                        flag = true;
+                }
+                if (!flag)
+                    file.delete();
             }
         });
         create();
@@ -88,21 +97,27 @@ public class Data {
      * @param identifier  Identifier of file
      */
     public static File cache(byte[] content, String identifier) {
-        if (!created) {
-            create();
-        }
         try {
-            File file = new File(System.getProperty("user.home") + "/." + StringHelper.id(Application.getTitle()) + "/cache/" + identifier);
-            if (content == null)
-                file.mkdir();
-            file.createNewFile();
-            if (content != null) {
-                OutputStream writer = new FileOutputStream(file);
-                writer.write(content);
-                writer.flush();
-                writer.close();
+            String home = System.getProperty("user.home");
+            if (home != null) {
+                File dir = new File(home + "/." + StringHelper.id(Application.getTitle()));
+                dir.mkdir();
+                dir.createNewFile();
+                File cacheDir = new File(dir.getPath() + "/cache");
+                cacheDir.mkdir();
+                cacheDir.createNewFile();
+                File file = new File(home + "/." + StringHelper.id(Application.getTitle()) + "/cache/" + identifier);
+                if (content == null)
+                    file.mkdir();
+                file.createNewFile();
+                if (content != null) {
+                    OutputStream writer = new FileOutputStream(file);
+                    writer.write(content);
+                    writer.flush();
+                    writer.close();
+                }
+                return file;
             }
-            return file;
         } catch (FileNotFoundException e) {
             e.printStackTrace();
         } catch (IOException e) {
@@ -118,6 +133,24 @@ public class Data {
             e.printStackTrace();
         }
         return null;
+    }
+
+    public static boolean delete(String identifier) {
+        File file = new File(System.getProperty("user.home") + "/." + StringHelper.id(Application.getTitle()) + "\\cache\\" + identifier);
+        if (file.exists()) {
+            file.delete();
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Files marked as storage will NOT be cleared as cache. See Data#clearCache()
+     *
+     * @param filePath  Path to file/folder
+     */
+    public static void markPathAsStorage(String filePath) {
+        markedStorage.add(filePath);
     }
 
     /**
@@ -165,8 +198,34 @@ public class Data {
             data = new HashMap<>();
             if (dataObjectMap != null) {
                 for (String key : dataObjectMap.keySet()) {
-                    data.put(key, gson.fromJson(dataObjectMap.get(key).object, Class.forName(dataObjectMap.get(key).objectClass)));
+                    try {
+                        data.put(key, gson.fromJson(dataObjectMap.get(key).object, Class.forName(dataObjectMap.get(key).objectClass)));
+                    } catch (ClassNotFoundException e) {
+                    }
                 }
+            }
+            int session = 0;
+            if (has("session")) {
+                session = (int) get("session");
+                if (session < 0) {
+                    session = 0;
+                }
+            }
+            if (!has("app_version") || !((String) get("app_version")).equalsIgnoreCase(Application.getVersion())) {
+                for (File subFile : dir.listFiles()) {
+                    if (!subFile.isDirectory()) {
+                        subFile.delete();
+                    }
+                }
+                FileHelper.forEachSubFile(cacheDir, new FileIteration() {
+                    @Override
+                    public void iterate(File file) {
+                        file.delete();
+                    }
+                });
+                store("app_version", Application.getVersion());
+                Application.restart();
+                return session;
             }
             Local local = (Local) get("local");
             Application.setLocal(local != null ? local : Local.EN_US);
@@ -178,14 +237,8 @@ public class Data {
             if (theme != null) {
                 Application.switchTheme(theme);
             }
-            if (has("session")) {
-                int session = (int) get("session");
-                if (session < 0) {
-                    session = 0;
-                }
-                return session;
-            }
-        } catch (IOException | ClassNotFoundException e) {
+            return session;
+        } catch (IOException e) {
             e.printStackTrace();
         }
         return 0;
@@ -197,19 +250,23 @@ public class Data {
             store("prev_dir", FileDialog.PREVIOUS_DIRECTORY.getPath());
             store("color_theme", Application.getTheme().getColorTheme());
             store("session", Application.getSession() + 1);
-            File dir = new File(System.getProperty("user.home") + "/." + StringHelper.id(Application.getTitle()));
-            dir.mkdir();
-            dir.createNewFile();
-            File dataFile = new File(dir.getPath() + "/data.json");
-            dataFile.createNewFile();
-            Map<String, DataObject> dataObjectMap = new HashMap<>();
-            for (String key : data.keySet()) {
-                dataObjectMap.put(key, new DataObject(data.get(key)));
+            store("app_version", Application.getVersion());
+            String home = System.getProperty("user.home");
+            if (home != null) {
+                File dir = new File(home + "/." + StringHelper.id(Application.getTitle()));
+                dir.mkdir();
+                dir.createNewFile();
+                File dataFile = new File(dir.getPath() + "/data.json");
+                dataFile.createNewFile();
+                Map<String, DataObject> dataObjectMap = new HashMap<>();
+                for (String key : data.keySet()) {
+                    dataObjectMap.put(key, new DataObject(data.get(key)));
+                }
+                FileWriter writer = new FileWriter(dataFile);
+                writer.write(gson.toJson(dataObjectMap));
+                writer.flush();
+                writer.close();
             }
-            FileWriter writer = new FileWriter(dataFile);
-            writer.write(gson.toJson(dataObjectMap));
-            writer.flush();
-            writer.close();
         } catch (IOException e) {
             e.printStackTrace();
         }
